complete project!!!!!

+++ Server/IProductService.cs
@@ -5,6 +5,6 @@ namespace Service
 {
     public interface IProductService
     {
-        Task<IEnumerable<Product>> getProducts(string? desc, int? minPrice, int? maxPrice, int?[] categoryIds);
+        Task<IEnumerable<Product>> getProducts(string? desc, decimal? minPrice, decimal? maxPrice, int?[] categoryIds);
     }
 }
