+++ Server/ProductService.cs
@@ -14,7 +14,7 @@ namespace Service
         {
             _productRepository = productRepository;
         }
-       public async Task<IEnumerable<Product>> getProducts(string? name, int? minPrice, int? maxPrice, int?[] categoryIds)
+       public async Task<IEnumerable<Product>> getProducts(string? name, decimal? minPrice, decimal? maxPrice, int?[] categoryIds)
         {
             return await _productRepository.getProducts(name,minPrice,maxPrice,categoryIds);
         }
