+++ Z_Repository/IProductRepository.cs
@@ -4,7 +4,7 @@ namespace Repository
 {
     public interface IProductRepository
     {
-        Task<IEnumerable<Product>> getProducts(string? desc, int? minPrice, int? maxPrice, int?[] categoryIds);
+        Task<IEnumerable<Product>> getProducts(string? desc, decimal? minPrice, decimal? maxPrice, int?[] categoryIds);
         Task<Product> getProductById(int id);
 
     }
